관리자 고객상세 응답 완성

+++ src/main/java/com/meongnyang/shop/dto/request/admin/ReqModifyProductDto.java
@@ -24,7 +24,7 @@ public class ReqModifyProductDto {
     private Long categoryId;
     @NotNull
     private Long productPrice;
-    private int productPriceDiscount;
+    private Long productPriceDiscount;
     private String productDetail;
     private String productBrand;
     private String productModel;
