+++ src/main/java/com/meongnyang/shop/dto/request/admin/ReqRegisterProductDto.java
@@ -20,7 +20,7 @@ public class ReqRegisterProductDto {
     private Long categoryId;
     @NotNull
     private Long productPrice;
-    private int productPriceDiscount;
+    private Long productPriceDiscount;
     private String productDetail;
     private String productBrand;
     private String productModel;
