+++ src/main/java/com/meongnyang/shop/dto/response/user/RespProductAllDto.java
@@ -19,7 +19,7 @@ public class RespProductAllDto {
         private String petGroupName;
         private String categoryName;
         private Long productPrice;
-        private int productPriceDiscount;
+        private Long productPriceDiscount;
         private String productDetail;
         private String productBrand;
         private String productModel;
